# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: WeddingOS Unified API
  version: 1.0.0
  description: |
    Unified OpenAPI 3.0 specification for the WeddingOS microservices platform.

    This document was assembled from the actual service route files and related
    Zod / Prisma / Pydantic definitions in the monorepo. Paths are documented as
    gateway-facing routes using the service prefixes requested by the user
    (`/auth`, `/users`, `/vendors`, `/bookings`, `/payments`, `/execution`,
    `/notifications`, `/reviews`, `/search`, `/media`, `/chat`, `/ai`).

    Notes:
    - Most services return a common JSON envelope with `success`, `data`, and often `meta`.
    - Internal routes may accept either a bearer token with admin role or `x-internal-api-key`.
    - Chat real-time Socket.IO events are described in tag descriptions and README, while this spec focuses on HTTP endpoints.
servers:
  - url: http://localhost:8000
    description: Local API gateway
  - url: https://api-staging.weddingos.in
    description: Staging
  - url: https://api.weddingos.in
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /bookings:
    post:
      tags: [Bookings]
      summary: Create enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEnquiryRequest'
      responses:
        '201':
          description: Created booking enquiry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingEnvelope'
    get:
      tags: [Bookings]
      summary: List my bookings
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/BookingStatus'
      responses:
        '200':
          description: Booking list for customer or vendor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListEnvelope'
  /bookings/{id}:
    get:
      tags: [Bookings]
      summary: Get booking by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Booking details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /bookings/{id}/quote:
    post:
      tags: [Bookings]
      summary: Send quote for enquiry
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendQuoteRequest'
      responses:
        '200':
          description: Updated booking with quote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingEnvelope'
  /bookings/{id}/accept-quote:
    post:
      tags: [Bookings]
      summary: Accept vendor quote
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Quote accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingEnvelope'
  /bookings/{id}/cancel:
    post:
      tags: [Bookings]
      summary: Cancel booking
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelBookingRequest'
      responses:
        '200':
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingEnvelope'
  /bookings/internal/{id}/confirm:
    post:
      tags: [Bookings]
      summary: Confirm booking after payment
      description: Internal endpoint. Code comments note it should be protected by network policy rather than normal JWT auth.
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paymentId]
              properties:
                paymentId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingEnvelope'
  /bookings/admin/list:
    get:
      tags: [Bookings]
      summary: Admin list bookings
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit100'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/BookingStatus'
      responses:
        '200':
          description: Paginated bookings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBookingListEnvelope'
...
  /bookings/admin/top-vendors:
    get:
      tags: [Bookings]
      summary: Top vendors by confirmed/completed bookings
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 20, default: 5 }
      responses:
        '200':
          description: Top vendor breakdown
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      topVendors:
                        type: array
                        items:
                          type: object
                          properties:
                            vendorId: { type: string, format: uuid }
                            bookingCount: { type: integer }
                            revenuePaise: { type: integer }
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
...
  /bookings/{id}/vendor-note:
    patch:
      tags: [Bookings]
      summary: Save vendor note on booking
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorNoteRequest'
      responses:
        '200':
          description: Updated booking note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /bookings/{id}/complete:
    post:
      tags: [Bookings]
      summary: Mark booking completed
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Completed booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalApiKey:
      type: apiKey
      in: header
      name: x-internal-api-key
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit100:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ApiMeta:
      type: object
      properties:
        requestId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
    ErrorObject:
      type: object
      properties:
        code:
          type: string
          nullable: true
        message:
          type: string
        field:
          type: string
          nullable: true
        details:
          type: object
          additionalProperties: true
          nullable: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/ApiMeta'
...
    BookingStatus:
      type: string
      enum: [ENQUIRY, QUOTE_SENT, QUOTE_ACCEPTED, ADVANCE_PENDING, ADVANCE_PAID, CONFIRMED, CHECKIN, COMPLETED, CANCELLED_BY_CUSTOMER, CANCELLED_BY_VENDOR, DISPUTED, REFUNDED]
    EventType:
      type: string
      enum: [WEDDING_CEREMONY, RECEPTION, ENGAGEMENT, HALDI, MEHNDI, SANGEET, BACHELOR_PARTY, PRE_WEDDING_SHOOT, BIRTHDAY, ANNIVERSARY, CORPORATE, OTHER]
...
    CreateEnquiryRequest:
      type: object
      required: [vendorId, eventDate, eventType, eventCity]
      properties:
        vendorId: { type: string, format: uuid }
        packageId: { type: string, format: uuid }
        eventDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        eventType:
          type: string
        eventCity: { type: string, maxLength: 100 }
        requirements: { type: string, maxLength: 2000 }
        guestCount: { type: integer, minimum: 1, maximum: 100000 }
        specialNotes: { type: string, maxLength: 500 }
    Booking:
      type: object
      properties:
        id: { type: string, format: uuid }
        bookingNumber: { type: string, nullable: true }
        customerId: { type: string, format: uuid }
        vendorId: { type: string, format: uuid }
        packageId: { type: string, format: uuid, nullable: true }
        status:
          $ref: '#/components/schemas/BookingStatus'
        eventDate: { type: string, format: date-time }
        eventType:
          $ref: '#/components/schemas/EventType'
        eventCity: { type: string }
        requirements: { type: string, nullable: true }
        guestCount: { type: integer, nullable: true }
        specialNotes: { type: string, nullable: true }
        quotedAmountPaise: { type: integer, nullable: true }
        finalAmountPaise: { type: integer, nullable: true }
        vendorQuoteNote: { type: string, nullable: true }
        paymentId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    BookingEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            booking:
              $ref: '#/components/schemas/Booking'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    BookingListEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            bookings:
              type: array
              items:
                $ref: '#/components/schemas/Booking'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    SendQuoteRequest:
      type: object
      required: [quotedAmountPaise]
      properties:
        quotedAmountPaise: { type: integer, minimum: 10000 }
        note: { type: string, maxLength: 1000 }
    CancelBookingRequest:
      type: object
      properties:
        reason: { type: string, maxLength: 500 }
    VendorNoteRequest:
      type: object
      required: [note]
      properties:
        note: { type: string, maxLength: 1000 }
    AdminBookingListEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            bookings:
              type: array
              items:
                $ref: '#/components/schemas/Booking'
        meta:
          allOf:
            - $ref: '#/components/schemas/ApiMeta'
            - type: object
              properties:
                total: { type: integer }
                page: { type: integer }
                limit: { type: integer }
                pages: { type: integer }
...